requirements.txt
async-timeout==4.0.3
blinker==1.8.2
click==8.1.7
django-openid-auth==0.17
Flask==2.3.3
Flask-OpenID==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multiprocess==0.70.16
pydantic==2.8.2
PyYAML==6.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
talisker[gunicorn,gevent,flask,prometheus,raven]==0.21.3
valkey==6.0.0b1
Werkzeug==2.3.8